intial implmentetion of the Table Create endpoint

+++ tables/models.go
@@ -3,7 +3,7 @@ package tables
 // Table struct is a row record of the tables table in the database
 type Table struct {
 	ID          int    `json:"id" db:"id"`
-	ProjectID   int    `json:"project_id" db:"project_id"`
+	ProjectID   int64    `json:"project_id" db:"project_id"`
 	OID		  	string `json:"oid" db:"oid"`
 	Name 	  	string `json:"name" db:"name"`
 	Description string `json:"description" db:"description"`
